prove REFL and ASSUME correct
but first, change the meaning of sequents slightly: we are allowed to
assume the term valuation actually covers the free variables of the
conclusion. also, my statement of REFL_correct assumes that the term has
a semantics whereas John's only assumes it's well typed, since all his
terms have semantics.