add term_ok and type_ok relations to the definition of |-
the intention is to make the inference system only produce theorems
using pieces of previously produced types, terms, and theorems. for
theorems this was already guaranteed by the inductive relation
construction. so now we have two more inductive relations to guarantee
it for types and terms also. (the base cases are the primitive
constants.)