abandon de Bruijn syntax, prove VSUBST_simple_subst
the theorem says that VSUBST behaves the same as a simple subst
(non-capture-avoiding) under certain conditions (i.e. bound variable
names are all fresh enough).

the plan is to prove a similar theorem for INST (with a simple inst) and
then show it's always possible to create an alpha-equivalent term
satisfying the conditions.