make collect_labels_conv faster
by periodically defining a constant to hide the tail of the list (which
may be long), and only unfolding these constants when they bubble up to
the top level during the recursive calls.

this may or may not be enough for unlabelling the bootstrap code. might
also need to hide the long list from the EVAL of inst_labels, for
example.