yet another approach to code_labels_conv
don't precompute FLOOKUP theorems for the finite map, but use
flookup_fupdate_conv to compute each particular application of FLOOKUP
during the inst_labels on the fly. this is unlikely to be faster unless
the Labs are very sparse in the code (e.g. much fewer than the total
number of labels). I would think in fact there are many _more_ Labs,
unfortunately.