replCorrect now builds, but with 1 cheat
The problem is in how the compiler compiles the initial environment.
Right now, it uses an empty fmap of types, which is bad because the
compile_top_thm requires all of the top-level declarations to appear in
the fmap.  So one way to fix it would be to make an fmap with the types
for the initial environment.  However, that seems pointless because
bindings shouldn't actually be printed for the initial state (as far as
I can tell from the repl's semantics).