tweak stateless syntax, and add new term_ok rule
change the representation of a defined constant tag to just include the
name and the definition term (rather than the whole equation term).

need to add term_ok for defined constants separately, because the new
theorem might be an implicit definition that doesn't mention the
constant (so wouldn't get picked up by the term in theorem ==> term_ok
rule).