Tweaked the semantics of application
Before when an application needed to raise a Div or Eq exception, it
created a (raise Div) expression and returned the environment
of the division (or =) expression. This is not good since now the cenv
is more lexical and the enclosing one could have the expression
shadowed. Now it just returns an exn_env that binds only the built-in
exceptions.