Remove the add_infx of --> in ml_translatorTheory.
The problem is that --> is already in the term grammar as an infix at a
different precedence for quotient$-->. We depend on quotientTheory
because of integers. I'm not sure what the right thing to do is here,
but having two different rules for the same token causes warning
messages and seems to make the EmitTeX munger fail.

Luckily, it seems that not much depends on the exact precedence of
ml_translator's --> (Arrow).