various tweaks to the HOL syntax and semantics
Use finite maps for the theory signature, to avoid the issue of
duplicate definitions. context_ok checks that the list of context
updates has no duplicates; to extract a signature from this list it is
turned into a finite map.

Also, give type abbreviations for various things that seem important:
assignments, valuations, signatures, interpretations... and clean up the
treatment of interpretations slightly.