(bad state) partial switch to using RDA's polymorphic values
also several lemmas about extending/reducing context, which motivated
this change. however, it seems like polymorphic values are probably a
mistake: we don't want term valuations to be general over type
valuations, since we actually always know what type valuation is
relevant. so, will probably go back to making being a term valuation
depend on a particular type valuation.

the other motivation for the change was to make type valuations
unconstrained on bad types (so a type valuation for an extended
signature is still a type valuation for any reduction of that signature:
it's allowed to send the undefined stuff anywhere). will keep this part
of the change.