prove new_axiom correct (after fixing it)
easy because we're concerned about validity of sequents in context:
we're allowed to assume all the axioms in the context hold, so adding a
new one doesn't make anything unproveable and lets you prove it too.

of course, we will also need at some point a result saying that certain
contexts actually do have models (e.g. if they only declare as axioms
the particular axioms we know how to prove are true)