move some things around, start on new_basic_type_definition_thm
but it is, as stated, false, because of a mismatch between how I thought
context extensions are represented in the state and wht the
implementation is desigend for. specifically: the implementation makes
NewConst/NewType updates in addition to the ConstSpec/TypeDefn updates,
whereas I only expected the latter. not sure which one to change...