Some progress on IL3
The compilation to IL3 wants to create sequences of expressions to evaluate. At
the moment, you have to use Let x = e1 in e2 where x is not free in e2 to
sequence e1 and e, but I won't want to have to do the check. So I'll add a
sequencing form to CakeML, probably by making the the variable bound by the Let
be optional.