Build up a compset for evaluating prog_to_bytecode.
The various parts of this compset should be moved out into appropriate
libraries where they can be maintained and optimised. It is somewhat
laborious to put the compset together by hand, but the advantage is you
get complete control and inspection of what's in there, and can
experiment with turning evaluation of certain things on and off.