Move and update labels_compute machinery, use in progToBytecodeScript.sml
At the moment, label removal uses mk_thm to prove theorems of the form

  |- code_labels_ok [... bytecode ...]

which should instead be proved from some compiler correctness theorem,
presuming the bytecode came from the compiler, and added to the database
in labels_computeLib. Getting hold of the bytecode mid-computation might
be a bit tricky, but perhaps could be done by hooking a conversion onto
a marker constant.

Alternatively we could write a conversion to prove code_labels_ok
directly, but that doesn't seem worth the effort.