make exh_ctors_env use a set type rather than a list
but use finite maps from nat to unit rather than lem sets, since we want
to be able to translate them. this is a placeholder type until we start
using an efficient representation for sets of nats. (abstract out the
nat_set type and operations to facilitate a change later.)

proofs not updated.