strengthen statement of decs_to_i2_correct
have not yet proved the strengthened claimed, but have shown that it is
sufficient for prompt_to_i2_correct. the new conclusion says that if the
decs appear to have been from a module and we encounter an error, then
envC_tagged holds if we add the module constructor env anyway.