use eval_real_inst_length for code_labels_conv in cakeml_compute
labels_computeLib.code_labels_conv needs a conversion to evaluate the
length function. previously had been passing in the cakeml compset
evaluator, but this doesn't include evaluation of real_inst_length. it
actually probably makes more sense to use a conversion that is
specialised to evaluating real_inst_length, so that is what we do in
this commit. if it needs to become more versatile later (e.g. for other
instruction lengths) we could make eval_real_inst_length a reference.