prove COMPILER_RUN_INV_repl_step
with a cheat about nothing in the globals of the bootstrap_bc_state
being NONE; the invariants in the early languages aren't really set up
to prove this semantically. if we can't/don't want to prove this by
evaluation, then they should probably be strengthened (I'm thinking of
global_env_inv).