Refactor the printer a bit. Broken proofs.
The repl was converting types to strings too soon, so that we had
branches on whether it was about to print the string "<word8>" or not.
This refactors so that the string conversion happens at the last minute.
This had the effect of moving the inf_type_to_string function from
repl_funScript.sml to compile.lem and thus needed some refactoring to
get the definition of infer_t into Lem.

As long as we're at it, we now traverse the list of types and bindings
together for printing, rather than turning the types into a finite map
and then traversing the bindings and looking them up in the map.