Start implementing 'a vector.
metatheory works, but has 3 cheats. The inferencer has been updated, but
the proof doesn't build. The compiler and translator have not been
updated.

Surprisingly, so far most of the work has been because the fromList
primitive takes in a list, which is not itself a primitive type (it
starts in the initial environment instead). But future primitives that
involve lists (e.g., for arrays or strings) should also benefit from
this work.