reduce COMPILER_RUN_INV_INL to a similar cheat
as COMPILER_RUN_INV_INR, and work on that cheat a little. it looks like
it's just going to be a huge pain to prove, although it's obviously
true. there are probably smarter less painful ways to prove it, somehow.

Basically, the cheat is: if a value satisfies the refinement invariant
for INPUT_TYPE then its refinement at intLang doesn't contain any
labels (which only appear on closures).