Start on refactoring the CakeML compset infrastructure
In each directory dirname, there should be a file compute_dirnameLib.sml
It should contain functions add_whatever_compset that adds theorems and
conversion relevant to the whatever to a given compset. It should also
contain a definition the_dirname_compset that creates a compset with
everything needed to run the various whatevers in dirname.

It doesn't work currently, because label removal in bytecode relies on
el_check which is defined in compiler, but the directories depend the
other way around.

The unify and label removal infrastructure had global state, I've
refactored so that the state is created locally to the add_..._compset
call.