Update the initial prog. with Word8Array
There is a problem with making primitive types usable in the program.
The type "Word8Array.array" should be there, but without type
abbreviations, there is no way to put it there and have it be the same
as the primitive type of arrays that all of the operations use. The
elaborator can be extended to handle this (just as it handles the
bindings for "int", "bool", etc), but that is less than ideal because
the initial environment doesn't exclusively come from the initial
program.